Fix issue where multiple paginators use same input
Previously it would try to add extensions to that shape multiple times. Fix creates a set of shapes that need extended and generates the extension code for these instead of using the array of operations that are paginated.
